More are_equal to vermouth.utils
